marks in markdown
Note that here we can see your version where the marks are as you
intended, but then it gets changed immediately when I open it to show
different marks for degree, accents, etc. TURNS OUT we have to use
special indicators for these markers, as long as we slowly learn all of
these will be OK. I didn’t check how to do accent yet.